Finished up general layout for level select screen
Put all the levels buttons in their appropriate places. (I set it up with the level layout excluding the elite and time bomb enemies, incase  Henry couldn't make them.) I used squares to connect the levels and show the player what level is which. If someone knows a better way to do that, please either tell me how to do that or just do it yourself and let me know that you did that. Please and thank you.