Organized the enemy paths in folder based on the level rather then the type of enemies
Lost one pathway for level 7 doing this, so instead I just replaced the missing path with the first normal enemy path on level 7. If I do everything I need to do before we turn this in, if I have time for it, I will probably make a new second normal enemy path for level 7, but rn I can't be bothered with it.