Made a lightweight enemy path
I think this might be the only lightweight path for this level, just cause of how I set the pathway. I feel like if we added any more lightweights, on top of every other enemy it'd be too much. Idk you guys can look at it tomorrow and decide idk